Consolidate entry point script logic.
